% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_pv_bfsnr_values}
\alias{patch_pv_bfsnr_values}
\title{Patch BFSNR Values for PV Installations}
\usage{
patch_pv_bfsnr_values(con, bfsnr_to, mapping = NULL)
}
\arguments{
\item{con}{A valid database connection object.}

\item{bfsnr_to}{A numeric scalar value of the new BFSNR code to apply.}

\item{mapping}{A data.frame specifying the street names and number whose
BFSNR needs to be updated. \code{NULL} by default.}

\item{name}{A string specifying the table name to update.}
}
\value{
The database connection, invisibly.
}
\description{
Updates the BFSNR (municipality code) values for specific
addresses in the photovoltaic installations database table:
\code{strom_produktionsanlagen}.
}
\details{
This function updates the municipality code (BFSNR) for a predefined
set of addresses in the photovoltaic installations table. It executes each
update within a transaction for safety, rolling back changes if errors
occur.

The function handles the special case of Thorenbergstrasse which has no
street number by looking for rows where the street number is either NULL or
empty.
}
